Remove Grades v0 REST API (DEPR-3)
