+++ lms/djangoapps/grades/api/v1/urls.py
@@ -3,7 +3,6 @@ from django.conf import settings
 from django.conf.urls import url
 
 from lms.djangoapps.grades.api.v1 import gradebook_views, views
-from lms.djangoapps.grades.api.views import CourseGradingPolicy
 
 
 app_name = 'lms.djangoapps.grades'
@@ -21,7 +20,7 @@ urlpatterns = [
     ),
     url(
         r'^policy/courses/{course_id}/$'.format(course_id=settings.COURSE_ID_PATTERN),
-        CourseGradingPolicy.as_view(),
+        views.CourseGradingPolicy.as_view(),
         name='course_grading_policy'
     ),
     url(
