+++ lms/djangoapps/grades/api/v1/views.py
@@ -2,9 +2,15 @@
 import logging
 from contextlib import contextmanager
 
+from rest_framework import status
+from rest_framework.generics import ListAPIView
+from rest_framework.response import Response
+
 from edx_rest_framework_extensions import permissions
 from edx_rest_framework_extensions.auth.jwt.authentication import JwtAuthentication
 from edx_rest_framework_extensions.auth.session.authentication import SessionAuthenticationAllowInactiveUser
+from lms.djangoapps.courseware.access import has_access
+from lms.djangoapps.grades.api.serializers import GradingPolicySerializer
 from lms.djangoapps.grades.api.v1.utils import (
     CourseEnrollmentPagination,
     GradeViewMixin,
@@ -14,7 +20,9 @@ from lms.djangoapps.grades.api.v1.utils import (
 )
 from lms.djangoapps.grades.course_grade_factory import CourseGradeFactory
 from lms.djangoapps.grades.models import PersistentCourseGrade
+from opaque_keys import InvalidKeyError
 from openedx.core.lib.api.authentication import OAuth2AuthenticationAllowInactiveUser
+from xmodule.modulestore.django import modulestore
 
 log = logging.getLogger(__name__)
 
@@ -136,3 +144,68 @@ class CourseGradesView(GradeViewMixin, PaginatedAPIView):
                     user_grades.append(self._serialize_user_grade(user, course_key, course_grade))
 
         return self.get_paginated_response(user_grades)
+
+
+class CourseGradingPolicy(GradeViewMixin, ListAPIView):
+    """
+    **Use Case**
+
+        Get the course grading policy.
+
+    **Example requests**:
+
+        GET /api/grades/v1/policy/courses/{course_id}/
+
+    **Response Values**
+
+        * assignment_type: The type of the assignment, as configured by course
+          staff. For example, course staff might make the assignment types Homework,
+          Quiz, and Exam.
+
+        * count: The number of assignments of the type.
+
+        * dropped: Number of assignments of the type that are dropped.
+
+        * weight: The weight, or effect, of the assignment type on the learner's
+          final grade.
+    """
+    allow_empty = False
+
+    authentication_classes = (
+        JwtAuthentication,
+        OAuth2AuthenticationAllowInactiveUser,
+        SessionAuthenticationAllowInactiveUser,
+    )
+
+    def _get_course(self, request, course_id):
+        """
+        Returns the course after parsing the id, checking access, and checking existence.
+        """
+        try:
+            course_key = get_course_key(request, course_id)
+        except InvalidKeyError:
+            raise self.api_error(
+                status_code=status.HTTP_400_BAD_REQUEST,
+                developer_message='The provided course key cannot be parsed.',
+                error_code='invalid_course_key'
+            )
+
+        if not has_access(request.user, 'staff', course_key):
+            raise self.api_error(
+                status_code=status.HTTP_403_FORBIDDEN,
+                developer_message='The course does not exist.',
+                error_code='user_or_course_does_not_exist',
+            )
+
+        course = modulestore().get_course(course_key, depth=0)
+        if not course:
+            raise self.api_error(
+                status_code=status.HTTP_404_NOT_FOUND,
+                developer_message='The course does not exist.',
+                error_code='user_or_course_does_not_exist',
+            )
+        return course
+
+    def get(self, request, course_id, *args, **kwargs):  # pylint: disable=arguments-differ
+        course = self._get_course(request, course_id)
+        return Response(GradingPolicySerializer(course.raw_grader, many=True).data)
